# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is api document.",
        "title": "Niveau API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "paths": {
...
        "/categories": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieve a paginated list of categories with optional search and sorting criteria.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories Modules"
                ],
                "summary": "Get list of categories",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by category name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "active",
                            "deactive"
                        ],
                        "type": "string",
                        "example": "\"active\"",
                        "description": "Filter by Status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "name",
                            "created_at"
                        ],
                        "type": "string",
                        "default": "created_at",
                        "example": "\"name\"",
                        "description": "Sort by field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "example": "\"asc\"",
                        "description": "Sort order",
                        "name": "sort_order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "Page number (must be \u003e= 1)",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "example": 10,
                        "description": "Items per page (between 10-100)",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with a list of categories",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_quydmfl_niveau-test_api_v1.SearchCategoryResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new category with the given details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories Modules"
                ],
                "summary": "Create a new category",
                "parameters": [
                    {
                        "description": "Category creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_quydmfl_niveau-test_api_v1.CreateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Category created successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_quydmfl_niveau-test_api_v1.Response"
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieve detailed information about a category by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories Modules"
                ],
                "summary": "Get category details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/github_com_quydmfl_niveau-test_api_v1.GetCategoryDetailResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_quydmfl_niveau-test_api_v1.CreateCategoryRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "Crunchy Munch"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "active",
                        "deactive"
                    ],
                    "example": "active"
                }
            }
        },
...
        "github_com_quydmfl_niveau-test_api_v1.GetCategoryDetailData": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-01-01"
                },
                "id": {
                    "type": "string",
                    "example": "67389c69-2e78-413a-9d77-6b749520b127"
                },
                "name": {
                    "type": "string",
                    "example": "Food"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-01-01"
                }
            }
        },
        "github_com_quydmfl_niveau-test_api_v1.GetCategoryDetailResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/github_com_quydmfl_niveau-test_api_v1.GetCategoryDetailData"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_quydmfl_niveau-test_api_v1.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_quydmfl_niveau-test_api_v1.SearchCategoryResponse": {
            "type": "object",
            "required": [
                "page",
                "size"
            ],
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "page": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "size": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 10,
                    "example": 10
                },
                "total_pages": {
                    "type": "integer",
                    "example": 10
                },
                "total_rows": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}